pyproject.toml
[project]
name = "njucaptcha"
version = "0.1.0"
description = "南大统一身份认证验证码（数据集、识别模型、识别服务）"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.5",
    "lxml>=6.0.0",
    "numpy>=2.3.1",
    "onnx>=1.18.0",
    "onnxruntime>=1.22.1",
    "pillow>=11.3.0",
    "pycryptodome>=3.23.0",
    "requests>=2.32.4",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128" },
]
torchvision = [
  { index = "pytorch-cu128" },
]

[[tool.uv.index]]
url = "https://mirror.nju.edu.cn/pypi/web/simple"
default = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
